refactor(boot.js): encapsulate logic for easier testing
